Alpha 0.5 Release
First Alpha Release. This is not a branch of IllusionProgressBar project
but is totally rewrited with a new idea and a new more powerfull e
lightweight engine, with many effects, many options to customize the bar
and much more. You can also extend a particular class to create your own
gradient (instruction in near future).
This is an Alpha with many bug and few instructions then be carefull!